+++ kore/src/Kore/Attribute/Pattern.hs
@@ -24,7 +24,6 @@ import Data.Hashable
 import qualified Generics.SOP as SOP
 import qualified GHC.Generics as GHC
 
-import Kore.Attribute.Pattern.ConstructorLikeTop
 import Kore.Attribute.Pattern.Created
 import Kore.Attribute.Pattern.Defined
 import Kore.Attribute.Pattern.FreeVariables
@@ -50,7 +49,6 @@ data Pattern variable =
         -- ^ The free variables of the pattern.
         , functional :: !Functional
         , function :: !Function
-        , constructorLikeTop :: !ConstructorLikeTop
         , defined :: !Defined
         , created :: !Created
         , simplified :: !Simplified
@@ -74,7 +72,6 @@ instance
     , Synthetic (FreeVariables variable) base
     , Synthetic Functional base
     , Synthetic Function base
-    , Synthetic ConstructorLikeTop base
     , Synthetic Defined base
     , Synthetic Simplified base
     ) =>
@@ -86,7 +83,6 @@ instance
             , freeVariables = synthetic (freeVariables <$> base)
             , functional = synthetic (functional <$> base)
             , function = synthetic (function <$> base)
-            , constructorLikeTop = synthetic (constructorLikeTop <$> base)
             , defined = synthetic (defined <$> base)
             , created = synthetic (created <$> base)
             , simplified = synthetic (simplified <$> base)
